transcription-interpretation-system/requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.1
openai>=1.35.0
anthropic>=0.34.0
google-generativeai>=0.8.2
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
mediapipe>=0.10.14
tensorflow-macos>=2.15; sys_platform == "darwin"
tensorflow-metal>=1.1; sys_platform == "darwin"


